requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
flake8==4.0.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
